+++ lib/eticon_api.dart
@@ -1,3 +1,5 @@
+///Eticon API is a handy and simple package for handling http requests.
+
 library eticon_api;
 
 export 'src/eticon_api.dart';